Moved React components to separate files in new directories
