+++ ui/jstests/util/test-select2.jsx
@@ -1,8 +1,6 @@
-define(['QUnit', 'jquery', 'utils', 'react', 'test_utils'],
-  function(QUnit, $, Utils, React, TestUtils) {
+define(['QUnit', 'jquery', 'select2_component', 'react', 'test_utils'],
+  function(QUnit, $, Select2, React, TestUtils) {
     'use strict';
-
-    var Select2 = Utils.Select2;
 
     var options = [
       {id: 'one', text: 'One'},
