+++ ui/static/ui/js/learningresources/term_list_item.jsx
@@ -1,0 +1,15 @@
+define("term_list_item", ["react"],
+  function (React) {
+    'use strict';
+
+    return React.createClass({
+      render: function () {
+        return (
+          <li className="applied-term list-group-item">
+            <strong>{this.props.label}: </strong> {this.props.terms}
+          </li>
+        );
+      }
+    });
+  }
+);
