+++ ui/static/ui/js/learningresources/xml_panel.jsx
