+++ ui/static/ui/js/listing/lore_data_table.jsx
