+++ ui/static/ui/js/require_config.js
@@ -18,15 +18,34 @@ var REQUIRE_PATHS = {
   historyadapter: "history.js/scripts/compressed/history.adapter.jquery",
   spin: "spin.js/spin.min",
   csrf: "../ui/js/csrf",
-  listing: "../ui/js/listing.jsx?noext",
-  manage_taxonomies: "../ui/js/manage_taxonomies.jsx?noext",
-  xml_panel: "../ui/js/xml_panel.jsx?noext",
-  learning_resources: "../ui/js/learning_resources.jsx?noext",
-  listing_resources: "../ui/js/listing_resources.jsx?noext",
-  static_assets: "../ui/js/static_assets.jsx?noext",
-  lr_exports: "../ui/js/lr_exports.jsx?noext",
-  pagination: "../ui/js/pagination.jsx?noext",
-  utils: "../ui/js/utils.jsx?noext",
+
+  listing: "../ui/js/listing/listing.jsx?noext",
+  listing_resources: "../ui/js/listing/listing_resources.jsx?noext",
+  pagination: "../ui/js/listing/pagination.jsx?noext",
+
+  manage_taxonomies: "../ui/js/taxonomy/manage_taxonomies.jsx?noext",
+
+  learning_resources: "../ui/js/learningresources/learning_resources.jsx?noext",
+  learning_resource_panel: "../ui/js/learningresources/" +
+    "learning_resource_panel.jsx?noext",
+  static_assets: "../ui/js/learningresources/static_assets.jsx?noext",
+  term_list: "../ui/js/learningresources/term_list.jsx?noext",
+  vocab_select: "../ui/js/learningresources/vocab_select.jsx?noext",
+  term_select: "../ui/js/learningresources/term_select.jsx?noext",
+  term_list_item: "../ui/js/learningresources/term_list_item.jsx?noext",
+  xml_panel: "../ui/js/learningresources/xml_panel.jsx?noext",
+
+  lr_exports: "../ui/js/exports/lr_exports.jsx?noext",
+  exports_component: "../ui/js/exports/exports_component.jsx?noext",
+  exports_header: "../ui/js/exports/exports_header.jsx?noext",
+
+  utils: "../ui/js/util/utils.jsx?noext",
+  confirmation_dialog: "../ui/js/util/confirmation_dialog.jsx?noext",
+  icheckbox: "../ui/js/util/icheckbox.jsx?noext",
+  infinite_list: "../ui/js/util/infinite_list.jsx?noext",
+  react_overlay_loader: "../ui/js/util/react_overlay_loader.jsx?noext",
+  select2_component: "../ui/js/util/select2_component.jsx?noext",
+  status_box: "../ui/js/util/status_box.jsx?noext",
 };
 var SHIMS = {
   "icheck": {"deps": ["jquery"]},
