+++ ui/templates/repository.html
@@ -42,31 +42,70 @@
 {{ block.super }}
 {% compress js %}
 <script type="text/jsx"
-        src="{% static "ui/js/manage_taxonomies.jsx" %}">
+        src="{% static "ui/js/taxonomy/manage_taxonomies.jsx" %}">
 </script>
 <script type="text/jsx"
-        src="{% static "ui/js/learning_resources.jsx" %}">
+        src="{% static "ui/js/learningresources/learning_resource_panel.jsx" %}">
 </script>
 <script type="text/jsx"
-        src="{% static "ui/js/static_assets.jsx" %}">
+        src="{% static "ui/js/learningresources/learning_resources.jsx" %}">
 </script>
 <script type="text/jsx"
-        src="{% static "ui/js/lr_exports.jsx" %}">
+        src="{% static "ui/js/learningresources/static_assets.jsx" %}">
 </script>
 <script type="text/jsx"
-        src="{% static "ui/js/listing_resources.jsx" %}">
+        src="{% static "ui/js/learningresources/term_list.jsx" %}">
 </script>
 <script type="text/jsx"
-        src="{% static "ui/js/xml_panel.jsx" %}">
+        src="{% static "ui/js/learningresources/term_list_item.jsx" %}">
 </script>
 <script type="text/jsx"
-        src="{% static "ui/js/pagination.jsx" %}">
+        src="{% static "ui/js/learningresources/term_select.jsx" %}">
 </script>
 <script type="text/jsx"
-        src="{% static "ui/js/utils.jsx" %}">
+        src="{% static "ui/js/learningresources/vocab_select.jsx" %}">
 </script>
 <script type="text/jsx"
-        src="{% static "ui/js/listing.jsx" %}">
+        src="{% static "ui/js/learningresources/xml_panel.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/exports/lr_exports.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/exports/exports_component.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/exports/exports_header.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/listing/listing_resources.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/listing/listing.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/listing/pagination.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/util/utils.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/util/confirmation_dialog.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/util/icheckbox.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/util/infinite_list.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/util/react_overlay_loader.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/util/select2_component.jsx" %}">
+</script>
+<script type="text/jsx"
+        src="{% static "ui/js/util/status_box.jsx" %}">
 </script>
 {% endcompress %}
 <script type="text/javascript">
